src/backend/requirements.txt
numpy
pandas
requests
scikit-learn
pydantic
uvicorn
fastapi
passlib[bcrypt]
python-dotenv
openai
pinecone-client
sqlalchemy
duckdb
duckdb-engine
cloud-sql-python-connector[pg8000]
